+++ bin/config/global_config.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 # Common global variables / constants
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 __tool_name_long__ = "Akamai Unified Log Streamer"
 __tool_name_short__ = "ULS"
 
@@ -9,9 +9,14 @@ __tool_name_short__ = "ULS"
 # Generic config
 bin_python = "python3"                          # Python binary to use (use OS standard when not using path)
 output_line_breaker = '\r\n'                    # Line breaking type (to split messages when streaming data)
+main_wait_default = 0.01                        # Default wait time within the main loop
+main_wait_max = 60                              # Maximum wait time for the main loop
+
     # EAA
-bin_eaa_cli = "ext/cli-eaa/bin/akamai-eaa"      # Path to the EAA CLI Executable
-eaa_cli_feeds = ['ACCESS', 'ADMIN']             # Available EAA CLI feeds
+        # Path to the EAA CLI Executable
+bin_eaa_cli = "ext/cli-eaa/bin/akamai-eaa"
+        # Available EAA CLI feeds
+eaa_cli_feeds = ['ACCESS', 'ADMIN', 'CONHEALTH']
     # ETP
 bin_etp_cli = "ext/cli-etp/bin/akamai-etp"      # Path to the ETP CLI Executable
 etp_cli_feeds = ['THREAT', 'AUP']               # Available ETP CLI feeds