+++ docs/CHANGELOG.md
@@ -1,4 +1,20 @@
 # Version History
+## v0.0.4
+|||
+|---|---|
+|Date|2021-06-17
+|Kind|Bugfix / Feature
+|Author|mschiess@akamai.com
+- Minor improvements 
+  - Wait_time and wait_max shifted to config
+  - added -f flag as alternative to --flag
+  - fixed an exception that was introduced in v0.0.3
+  - bumped MFA -CLI to 0.0.5 in dockerfile
+  - added an additional debugging example
+- Feature:
+  - EAA CONNECTOR HEALTH (CONHEALTH) now available
+  - Preflight (forced) check for available cli's
+
 ## v0.0.3
 |||
 |---|---|
@@ -10,7 +26,6 @@
 - fixed a bug that caused Popen PIPE to hang in certain circumstances
 - bumped Dockerfile to newer CLI versions
 - introduced RAW output (send data to stdout)
-
 
 ## v0.0.2
 |||